clean up wanrings

+++ examples/thermal_cam/thermal_cam.ino
@@ -103,7 +103,7 @@ void loop() {
 
   for(int i=0; i<AMG88xx_PIXEL_ARRAY_SIZE; i++){
     uint8_t colorIndex = map(pixels[i], MINTEMP, MAXTEMP, 0, 255);
-    colorIndex = constrain(colorIndex, 0, 255);
+    colorIndex = (uint8_t)constrain((int16_t)colorIndex, (int16_t)0, (int16_t)255);
 
     //draw the pixels!
     tft.fillRect(displayPixelHeight * floor(i / 8), displayPixelWidth * (i % 8),
