+++ examples/thermal_cam_interpolate/thermal_cam_interpolate.ino
@@ -173,7 +173,7 @@ void drawpixels(float *p, uint8_t rows, uint8_t cols, uint8_t boxWidth, uint8_t 
       else colorTemp = val;
       
       uint8_t colorIndex = map(colorTemp, MINTEMP, MAXTEMP, 0, 255);
-      colorIndex = constrain(colorIndex, 0, 255);
+      colorIndex = (uint8_t)constrain((int16_t)colorIndex, (int16_t)0, (int16_t)255);
       //draw the pixels!
       uint16_t color;
       color = val * 2;
